+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [7.13.0](https://github.com/Arize-ai/arize/compare/arize-python-sdk/v7.12.1...arize-python-sdk/v7.13.0) (2024-03-28)
+
+
+### 🎁 New Features
+
+* Allow sending latent tags without actuals ([#29387](https://github.com/Arize-ai/arize/issues/29387)) ([4c91949](https://github.com/Arize-ai/arize/commit/4c91949bcf861ed98b079953d3210b5276119bca))
+
+
+### 🐛 Bug Fixes
+
+* use pandas items() vs deprecated iteritems() ([#29384](https://github.com/Arize-ai/arize/issues/29384)) ([0eb377e](https://github.com/Arize-ai/arize/commit/0eb377e6174a9a917ef1e4a579e6874128747a68))
+
 ## [7.12.1](https://github.com/Arize-ai/arize/compare/arize-python-sdk-v7.12.0...arize-python-sdk/v7.12.1) (2024-03-26)